requirements.txt
Flask>=3.1.2
python-dotenv>=1.1.1
google-generativeai
numpy
pandas
pathway
requests
gunicorn

